1. Simple Conversational Agent/backend/requirements.txt
langchain
langchain_experimental
openai
python-dotenv
langchain_google_genai
fastapi
fastapi[standard]